refactor(neon_framework,neon_rich_text): Move rich text rendering to neon_rich_text
