+++ packages/neon_framework/example/pubspec_overrides.yaml
@@ -1,4 +1,4 @@
-# melos_managed_dependency_overrides: account_repository,cookie_store,dashboard_app,dynamite_runtime,files_app,interceptor_http_client,neon_framework,neon_http_client,neon_lints,news_app,nextcloud,notes_app,notifications_app,notifications_push_repository,sort_box,talk_app
+# melos_managed_dependency_overrides: account_repository,cookie_store,dashboard_app,dynamite_runtime,files_app,interceptor_http_client,neon_framework,neon_http_client,neon_lints,neon_rich_text,news_app,nextcloud,notes_app,notifications_app,notifications_push_repository,sort_box,talk_app
 dependency_overrides:
   account_repository:
     path: ../packages/account_repository
@@ -18,6 +18,8 @@ dependency_overrides:
     path: ../packages/neon_http_client
   neon_lints:
     path: ../../neon_lints
+  neon_rich_text:
+    path: ../packages/neon_rich_text
   news_app:
     path: ../packages/news_app
   nextcloud:
