+++ packages/neon_framework/lib/widgets.dart
@@ -7,7 +7,6 @@ export 'package:neon_framework/src/widgets/image.dart' hide NeonImage;
 export 'package:neon_framework/src/widgets/linear_progress_indicator.dart';
 export 'package:neon_framework/src/widgets/list_view.dart';
 export 'package:neon_framework/src/widgets/relative_time.dart';
-export 'package:neon_framework/src/widgets/rich_text/rich_text.dart';
 export 'package:neon_framework/src/widgets/server_icon.dart';
 export 'package:neon_framework/src/widgets/user_avatar.dart' hide NeonUserStatusIndicator;
 export 'package:neon_framework/src/widgets/user_status_icon.dart';
