+++ packages/neon_framework/packages/neon_rich_text/lib/neon_rich_text.dart
@@ -1,0 +1,5 @@
+export 'src/rich_objects/deck_card.dart';
+export 'src/rich_objects/fallback.dart';
+export 'src/rich_objects/file.dart';
+export 'src/rich_objects/mention.dart';
+export 'src/rich_text.dart';
