+++ packages/neon_framework/packages/neon_rich_text/lib/src/rich_objects/file.dart
