+++ packages/neon_framework/packages/neon_rich_text/test/goldens/rich_text_object_mention_user_highlight.png
