+++ packages/neon_framework/packages/neon_rich_text/test/rich_text_test.dart
@@ -10,6 +10,7 @@ import 'package:neon_framework/testing.dart';
 import 'package:neon_framework/theme.dart';
 import 'package:neon_framework/utils.dart';
 import 'package:neon_framework/widgets.dart';
+import 'package:neon_rich_text/neon_rich_text.dart';
 import 'package:nextcloud/core.dart' as core;
 import 'package:provider/provider.dart';
 import 'package:rxdart/rxdart.dart';
