+++ packages/neon_framework/packages/notifications_app/test/notification_test.dart
@@ -5,6 +5,7 @@ import 'package:mocktail/mocktail.dart';
 import 'package:neon_framework/models.dart';
 import 'package:neon_framework/testing.dart';
 import 'package:neon_framework/widgets.dart';
+import 'package:neon_rich_text/neon_rich_text.dart';
 import 'package:nextcloud/notifications.dart' as notifications;
 import 'package:notifications_app/l10n/localizations.dart';
 import 'package:notifications_app/src/widgets/action.dart';
